Added marketing functionality

+++ Insurance-final-project/Dto/PaymentDto.cs
@@ -9,8 +9,8 @@ namespace Insurance_final_project.Dto
         public string PolicyName { get; set; }
 
         [Required(ErrorMessage = "Amount required!")]
-        [Range(1, long.MaxValue, ErrorMessage = "Amount must be greater than 0.")]
-        public long Amount { get; set; }
+        [Range(1, double.MaxValue, ErrorMessage = "Amount must be greater than 0.")]
+        public decimal Amount { get; set; }
 
         [Required(ErrorMessage = "SuccessUrl required!")]
         [Url(ErrorMessage = "Invalid URL format for SuccessUrl.")]
